LayoutTests/fast/notifications/notification-permisssion-deletable.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Tests that Notification.permission is deletable.");

shouldBeEqualToString("Notification.permission", "default");
shouldBeTrue("delete Notification.permission");
shouldBeUndefined("Notification.permission");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
