LayoutTests/fast/notifications/notification-permisssion-deletable-expected.txt

Tests that Notification.permission is deletable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Notification.permission is "default"
PASS delete Notification.permission is true
PASS Notification.permission is undefined.
PASS successfullyParsed is true

TEST COMPLETE
